% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{create_repo}
\alias{create_repo}
\alias{create_private_repo}
\title{Create Github repository}
\usage{
create_repo()

create_private_repo()
}
\value{
Github repo created for associated account.
}
\description{
Creates public repository on Github (from local repository)
}
